+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved clarity of settings panel error messages.
 - Improved performance.
 
 ## [1.5.0] - 2025-02-02